Extract handlers for watching server to a pkg
- The extracted handlers use in golang standard context, protobuf
  request/response messages.

- Hence, these can be shared between grpc and other upcoming
  implementations (like peruniotcp server).
